==> 02_pre_processing.Rmd <==
---
title: "02_pre_processing.Rmd"
author: "Leeya Pressburger"
date: "`r Sys.Date()`"
output:
  bookdown::html_document2:
    fig_caption: yes
    number_sections: yes
    toc: true
    toc_depth: 2
    toc_float: true
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

### R setup

Load necessary packages, read in an ini file and initiate core, and set seed for reproducibility.

```{r intro, warning=FALSE, message=FALSE}
library(hector)
library(dplyr)
library(tidyr)
library(ggplot2)
library(bookdown)
library(data.table) 

theme_set(theme_light())

source("trackingC_common.R")

# Read in data files, generated by trackingC_setup.Rmd
runlist <- fread("./output_files/runlist_test.csv")
model_output <- fread("./output_files/model_output_test.csv")
trk_output <- fread("./output_files/trk_output_test.csv")

# Add ssp information to the output data
model_output$ssp <- runlist$ssp[model_output$run_number]
trk_output$ssp <- runlist$ssp[trk_output$run_number]

```

### Hector atmospheric CO~2~ vs observed data

``` {r ca-comp, echo = FALSE, fig.cap = " Historical CO2 data from Hector runs and CMIP6 data."}
# This chunk will compare CMIP6 historical CO2 concentration data with 
# Hector data generated by our model runs

# First, we need to read in the CMIP6 data and organize
# Read in CMIP CO2 data, compute summary statistics
cmip_co2 <- read.csv("./data_files/CMIP6_annual_co2.csv")

# Isolate historical data
cmip_co2_hist <- cmip_co2 %>%
  filter(year %in% 1850:2014,
         experiment == "historical") %>%
  mutate(value = value * 1e6,
         units = "ppm") %>%
  group_by(year) %>%
  summarize(sdev = sd(value), 
            minimum = min(value),
            maximum = max(value),
            med = median(value),
            source = "CMIP6",
            .groups = "drop")

# Identify unrealistic runs - for each Hector run, if more than 33% of values are
# outside of the min/max of the CMIP6 data, the run is removed from the data

# Find min/max of the CMIP6 data
cmip_bounds_h <- cmip_co2_hist %>% 
  select(year, minimum, maximum) %>%
  ungroup()

# Get CO2 data from Hector, add "fail" column if values are outside the
# min/max of CMIP6
hector_co2 <- model_output %>%
  filter(variable == "CO2_concentration",
         year >= 1850 & year <= 2014) %>%
  select(run_number, year, value) %>%
  left_join(cmip_bounds_h, by = "year") %>%
  mutate(fail = value < minimum | value > maximum)

# Summarize
run_fail_summary_h <- hector_co2 %>%
  group_by(run_number) %>%
  summarize(fraction_fail = sum(fail) / n()) %>%
  filter(fraction_fail < 0.33)

# Filter out 'unrealistic' runs (identified above)
model_output_filtered_1 <- run_fail_summary_h %>%
  left_join(model_output, by = "run_number") %>% 
  select(-fraction_fail)

# Remove large data
rm(model_output)

# Note that the tracking output needs to be filtered as well
trk_output_filtered_1 <- run_fail_summary_h %>%
  left_join(trk_output, by = "run_number") %>% 
  select(-fraction_fail)

# Remove large data and force clear memory
rm(trk_output)
gc()

## Plot
# Read in filtered Hector data, compute summary statistics 
hector_co2_summary <- model_output_filtered_1 %>%
  filter(variable == "CO2_concentration",
         year %in% 1850:2014,
         ssp == MAIN_SCENARIO) %>%
  select(run_number, year, value) %>%
  mutate(units = "ppm") %>%
  group_by(year) %>%
  summarize(sdev = sd(value), 
            minimum = min(value),
            maximum = max(value),
            med = median(value),
            source = "Hector",
            .groups = "drop")

# Combine data
plot_data_co2 <- bind_rows(cmip_co2_hist, hector_co2_summary)

# Plot
co2_plot <- plot_data_co2 %>% 
  filter(year < 2015) %>%
  ggplot(aes(year, fill = source, color = source)) +
  geom_ribbon(aes(ymin = minimum, ymax = maximum), alpha = 0.15, color = NA) +
  geom_ribbon(aes(ymin = med - sdev, ymax = med + sdev), alpha = 0.5, color = NA) +
  geom_line(aes(y = med), size = 2) +
  scale_color_viridis_d(begin = 0.5) +
  scale_fill_viridis_d(begin = 0.5) +
  labs(x = "Year",
       y = "CO2 (ppm)",
       color = "Source",
       fill = "Source") +
  scale_x_continuous(expand = c(0, 0)) ; co2_plot

```


### Hector temperature vs. CMIP6 

``` {r filtered, warning = FALSE, message = FALSE, echo = FALSE, fig.cap = "Global temperature anomaly from CMIP6 and filtered Hector runs. Lines show median; shaded areas show ±1 s.d. of median (darker) and minimum and maximum of ensemble (lighter)."}
# This chunk will compare CMIP6 historical and projected global temperature
# anomaly data with Hector data generated by our model runs

# Read in CMIP6 comparison data
cmip_raw <- read.csv("./data_files/CMIP6_annual_tas_global.csv")

# Isolate historical data
cmip_tgav_hist <- cmip_raw %>% 
  filter(experiment == "historical",
         variable == "Tgav",
         year <= 2100) %>%
  group_by(year) %>%
  summarize(sdev = sd(value),
            minimum = min(value),
            maximum = max(value),
            med = median(value)) %>%
  mutate(source = "CMIP6") 

# Isolate desired scenarios
cmip_tgav_sce <- list()
for(n in names(SSP_files)){
  cmip_tgav_sce[[n]] <- cmip_raw %>% 
    filter(experiment == n)
}  

cmip_sce <- bind_rows(cmip_tgav_sce)
rm(cmip_raw)

# Filter for Tgav and calculate summary metrics
cmip_t_scenarios <- cmip_sce %>%
  rename(ssp = experiment) %>%
  filter(variable == "Tgav",
         year <= 2100) %>%
  group_by(year, ssp) %>%
  summarize(sdev = sd(value),
            minimum = min(value),
            maximum = max(value),
            med = median(value)) %>%
  mutate(source = "CMIP6") %>%
  arrange(ssp, year)

# Need to append the historical values to each SSP
# Add SSP column to historical data, repeat for all scenarios
cmip_hist_repeat_t <- list()
for(n in names(SSP_files)){
  cmip_hist_repeat_t[[n]] <- cmip_tgav_hist %>% 
    mutate(ssp = n)
}
cmip_hist_repeat <- bind_rows(cmip_hist_repeat_t)

cmip_t_final <- bind_rows(cmip_hist_repeat, cmip_t_scenarios) %>% 
  arrange(ssp, year)

### Now that we have all the data we'll need in one place, we want to filter out 
# run numbers twice.
# First, we just want to look at the historical data. Then we'll look at each SSP

# Identify unrealistic historical runs - for each run, if more than 33% of 
# values are outside of the min/max of the CMIP6 data, the run is dropped
cmip_bounds_t <- cmip_t_final %>% 
  group_by(ssp) %>%
  select(year, minimum, maximum) %>%
  ungroup()

# Get temperature data from Hector, add "fail" column if values are outside the
# min/max of CMIP6 runs
tgav_h <- model_output_filtered_1 %>%
  filter(variable == "global_tas",
         year >= 1850 & year <= 2014) %>%
  select(run_number, ssp, year, value) %>%
  left_join(cmip_bounds_t, by = c("year", "ssp")) %>%
  mutate(fail = value < minimum | value > maximum)

# Summarize
run_fail_summary_tgav <- tgav_h %>%
  group_by(ssp, run_number) %>%
  summarize(fraction_fail = sum(fail) / n()) %>%
  filter(fraction_fail < 0.33)

# Filter out 'unrealistic' runs (identified above)
model_output_filtered_2 <- run_fail_summary_tgav %>%
  left_join(model_output_filtered_1, by = c("ssp", "run_number")) %>% 
  select(-fraction_fail)

# Remove large data
rm(model_output_filtered_1)

# Note that the tracking output needs to be filtered as well
trk_output_filtered_2 <- run_fail_summary_tgav %>% 
  left_join(trk_output_filtered_1, by = c("ssp", "run_number")) %>% 
  select(-fraction_fail)

# Remove large data, force clear memory
rm(trk_output_filtered_1)
gc()

# Repeat for SSPs - future projections
# Identify unrealistic runs - for each run, if more than 50% of values are
# outside of the min/max of the CMIP6 data, the run is dropped

# Get temperature data from filtered Hector runs, add "fail" column if 
# values are outside the min/max of CMIP6
tgav_h_ssp <- model_output_filtered_2 %>%
  filter(variable == "global_tas",
         year >= 2015 & year <= 2100) %>%
  select(run_number, ssp, year, value) %>%
  left_join(cmip_bounds_t, by = c("year", "ssp")) %>%
  mutate(fail = value < minimum | value > maximum)

# Summarize
run_fail_summary_tgav_ssp <- tgav_h_ssp %>%
  group_by(ssp, run_number) %>%
  summarize(fraction_fail = sum(fail) / n()) %>%
  filter(fraction_fail < 0.5)

# Filter out 'unrealistic' runs (identified above)
model_output_final <- run_fail_summary_tgav_ssp  %>%
  left_join(model_output_filtered_2, by = c("ssp", "run_number")) %>% 
  filter(fraction_fail < 0.5)

# Remove large data
rm(model_output_filtered_2)

# Note that the tracking output needs to be filtered as well
trk_output_final <- run_fail_summary_tgav_ssp %>%
  left_join(trk_output_filtered_2, by = c("ssp", "run_number"))

# Remove large data, force clear memory
rm(trk_output_filtered_2)
gc()
  
# Create facet labels for plotting
ssp_labels <- c("SSP126", "SSP245", "SSP370", "SSP460", "SSP585")
names(ssp_labels) <- c(names(SSP_runs))

# Plot temperature spreads for each scenario, Hector vs CMIP6
temp_plot <- model_output_final %>%
  filter(variable == "global_tas",
         year >= 1850 & year <= 2100) %>%
  # compute median, min, max, etc. for all remaining runs
  group_by(year, ssp) %>%
  summarize(sdev = sd(value),
            minimum = min(value),
            maximum = max(value),
            med = median(value), .groups = "drop") %>% 
  # combine with CMIP6 data
  mutate(source = "Hector") %>%
  bind_rows(cmip_t_final) %>%
  # ...and plot
  ggplot(aes(year, fill = source, color = source)) +
  geom_ribbon(aes(ymin = minimum, ymax = maximum), alpha = 0.15, color = NA) +
  geom_ribbon(aes(ymin = med - sdev, ymax = med + sdev), alpha = 0.5, color = NA) +
  geom_line(aes(y = med), size = 1) +
  scale_color_viridis_d(begin = 0.5) +
  scale_fill_viridis_d(begin = 0.5) +
  facet_wrap(~ssp, scales = "free",
             labeller = labeller(ssp = ssp_labels),
             ncol = 3) +
  labs(x = "Year",
       y = expression(degree*C),
       color = "Source",
       fill = "Source") +
  theme(axis.text.x = element_text(angle = 90)) +
  scale_x_continuous(expand = c(0, 0)) ; temp_plot

```

## Hector temperature vs CMIP6 diagnostic

``` {r s-cmip, message = FALSE, warning = FALSE, fig.cap = " Visualization of sample model runs. The black dots represent the CMIP6 minimums and maximums for global temperature anomaly, and the runs that pass must fall within these bounds more than 50% of the time."}
# Visualize the run fail summaries calculated above
# Make a table of failure rates

# Historical CO2
run_fail_summary_h %>% 
  group_by(fraction_fail > 0.33) %>% 
  summarise(n()) %>% 
  knitr::kable()

# Historical temperature
run_fail_summary_tgav %>% 
  group_by(fraction_fail > 0.33) %>% 
  summarise(n()) %>% 
  knitr::kable()

# Future temperature
run_fail_summary_tgav_ssp %>% 
  group_by(fraction_fail > 0.5) %>% 
  summarise(n()) %>% 
  knitr::kable()

# ...and a figure 
hector_temp_runs <- tgav_h_ssp %>% 
  group_by(ssp) %>%
  filter(year == min(year)) %>%
  slice_sample(n = 100, replace = TRUE) %>% # this is a debugging figure; plot just 100 runs for clarity
  arrange(ssp, run_number) %>%
  summarize(run_number)

# Filter out data for just the 100 runs per scenario identified above
hector_temp_slice <- left_join(hector_temp_runs, tgav_h_ssp, by = c("run_number", "ssp"))

# Plot
left_join(hector_temp_slice, run_fail_summary_tgav_ssp, by = c("run_number", "ssp")) %>%
  ggplot(aes(year, value, color = fraction_fail, group = run_number)) + 
  geom_line() + 
  geom_point(aes(y = minimum), color = "black") + 
  geom_point(aes(y = maximum), color = "black") +
  facet_wrap(~ssp,
             labeller = labeller(ssp = ssp_labels),
             ncol = 3) +
  scale_color_viridis_c() +
  labs(x = "Year",
       y = expression(degree*C),
       color = "Fraction of failed runs")

```

``` {r cleanup-cmip, echo = FALSE}
# Write out data, clean environment
fwrite(model_output_final, "./output_files/model_output_final_test.csv", row.names = FALSE)
fwrite(trk_output_final, "./output_files/trk_output_final_test.csv", row.names = FALSE)

rm(list = ls())
gc()
```
